+++ typescript/web_core/src/v1_0/basic_catalog/index.ts
@@ -16,3 +16,4 @@
 
 export * from './components/basic_components.js';
 export * from './functions/basic_functions_api.js';
+export * from './functions/basic_functions.js';